requirements.txt
# Nolus Ambassador Dashboard Dependencies

# Core Flask
Flask>=3.0.0
python-dotenv>=1.0.0
Werkzeug>=3.0.0

# Web Scraping
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright-stealth>=2.0.1

# Google Sheets API (optional - for sync)
google-api-python-client>=2.111.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0

# Discord Bot
discord.py>=2.3.0

# Production Server
gunicorn>=21.0.0
